refact Server
