--- src/main/java/neointernship/chess/Main.java
@@ -1,23 +1,0 @@
-package neointernship.chess;
-
-/**
- * Главный класс приложения.
- */
-public class Main {
-    /**
-     * Главная функция, в которой мы инициализируем старт игры.
-     *
-     * @param args аргументы командной строки.
-     */
-    public static void main(final String[] args) {
-       /*
-       Selfplay selfplay = new Selfplay();
-        selfplay.start();
-        */
-
-
-
-        /*Selfplay selfplay = new Selfplay();
-        selfplay.start(10);*/
-    }
-}
