+++ src/main/java/neointernship/web/server/IServer.java
@@ -1,0 +1,5 @@
+package neointernship.web.server;
+
+public interface IServer {
+    void startServer();
+}
